+++ example/src/App.svelte
@@ -2,10 +2,10 @@
 	import Typewriter from 'Typewriter.svelte'
 </script>
 
-<Typewriter cascade on:done={() => console.log('Hello World!')}>
+<Typewriter loop on:done={() => console.log('Hello World!')}>
   <div>some text...</div>
-  <div>Foo</div>
-  <a href="user">some link</a>
-  <a href="user">yet another link</a>
-  <div>Bar</div>
+  <span>
+    <a href='user'>some link</a>
+    <a href='user'>yet another link</a>
+  </span>
 </Typewriter>