+++ example/src/App.svelte
@@ -3,9 +3,5 @@
 </script>
 
 <Typewriter loop on:done={() => console.log('Hello World!')}>
-  <div>some text...</div>
-  <span>
-    <a href='user'>some link</a>
-    <a href='user'>yet another link</a>
-  </span>
+  some & text...
 </Typewriter>