+++ CHANGELOG.md
@@ -9,12 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- fix(utils): escape ampersand unicode entity, closes #31 [`#31`](https://github.com/henriquehbr/svelte-typewriter/issues/31)
 - fix(utils): properly filter innerHTML tags, closes #32 [`#32`](https://github.com/henriquehbr/svelte-typewriter/issues/32)
 
 ### Commits
 
 - chore: bump dependencies [`46cf177`](https://github.com/henriquehbr/svelte-typewriter/commit/46cf17746734193a0d3b34a83f9ef703b5e63bda)
-- chore(example): output css to ./public/build [`aa13305`](https://github.com/henriquehbr/svelte-typewriter/commit/aa13305013cd81e07785ab27d4740ff21af964bc)
+- chore(example): output css to ./public/build [`898bbd4`](https://github.com/henriquehbr/svelte-typewriter/commit/898bbd4d45a10cdf2ca0af40ab06dd074c2f509d)
 
 ## [v2.4.3](https://github.com/henriquehbr/svelte-typewriter/compare/v2.4.2...v2.4.3) - 2020-12-11